refactor: rename sha256 package to hash
