+++ internal/hash/sha256/sha256.go
